Added world covid data by continents
