+++ src/components/CovidInfo/CovidWorld.js/CovidWorld.css
@@ -13,6 +13,7 @@
   flex-direction: column;
 }
 
+/* WORLD CARD CSS STARTS */
 .world_head {
   display: flex;
   flex-wrap: wrap;
@@ -49,6 +50,28 @@
   font-weight: 400;
 }
 
+/* WORLD CARD CSS ENDS */
+
+/* CONTINENTS CARD CSS STARTS */
+.continents_head {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  height: 122vh;
+  width: 100%;
+  margin-top: -30px;
+  font-family: monospace;
+}
+
+.count .count_heading {
+  font-size: 20px;
+  color: cadetblue;
+  display: flex;
+  justify-content: space-around;
+  font-weight: 700;
+  text-transform: uppercase;
+}
+
 @media screen and (max-width: 750px) {
   .world_head_paper h3 {
     font-size: 13px;
@@ -77,5 +100,25 @@
 
   .world_head_paper .count {
     font-size: 15px;
+  }
+
+  .continents_head {
+    height: 125vh;
+  }
+
+  .count .count_heading {
+    font-size: 29px;
+  }
+}
+
+@media screen and (max-width: 420px) {
+  .continents_head {
+    height: 115vh;
+  }
+}
+
+@media screen and (max-width: 360px) {
+  .continents_head {
+    height: 145vh;
   }
 }
