+++ src/components/CovidInfo/CovidWorld.js/CovidWorld.js
@@ -11,7 +11,7 @@ import "./CovidWorld.css";
 import {
   WorldPaperInformation,
   ContinentPaperInformation,
-} from "./PaperContents";
+} from "./CovidWorldContents";
 
 const CovidWorld = ({ value, index }) => {
   const [allWorldData, setAllWorldData] = useState([]);
@@ -20,7 +20,7 @@ const CovidWorld = ({ value, index }) => {
   // TODOs
   // 1. make a select field to filter out the slection
   //    - get whole world (done)
-  //    - get Data by continents
+  //    - get Data by continents (done)
   //    - get Data by specific continent
   //    - get Data by countries
   //    - get Data by country
@@ -28,9 +28,9 @@ const CovidWorld = ({ value, index }) => {
   const SelectOptions = [
     "Get COVID19 World Information",
     "Get COVID19 Data by continents",
-    "Get COVID19 Data by specific continent",
-    "Get COVID19 Data by countries",
-    "Get COVID19 Data by country",
+    // "Get COVID19 Data by specific continent",
+    // "Get COVID19 Data by countries",
+    // "Get COVID19 Data by country",
   ];
 
   const WorldPaperContents = [
@@ -116,89 +116,6 @@ const CovidWorld = ({ value, index }) => {
     },
   ];
 
-  const ContinentsPaperContent = [
-    {
-      paperTitle: "Total Active Cases",
-      paperAnswer: dataByContinent.active,
-    },
-    {
-      paperTitle: "Active per one million",
-      paperAnswer: dataByContinent.activePerOneMillion,
-    },
-    {
-      paperTitle: "Affected Countries",
-      paperAnswer: dataByContinent.affectedCountries,
-    },
-    {
-      paperTitle: "Total Cases",
-      paperAnswer: dataByContinent.cases,
-    },
-    {
-      paperTitle: "Active Cases per million",
-      paperAnswer: dataByContinent.casesPerOneMillion,
-    },
-    {
-      paperTitle: "critical",
-      paperAnswer: dataByContinent.critical,
-    },
-    {
-      paperTitle: "critical per million",
-      paperAnswer: dataByContinent.criticalPerOneMillion,
-    },
-    {
-      paperTitle: "deaths",
-      paperAnswer: dataByContinent.deaths,
-    },
-    {
-      paperTitle: "deaths Per OneMillion",
-      paperAnswer: dataByContinent.deathsPerOneMillion,
-    },
-    {
-      paperTitle: "one Case Per People",
-      paperAnswer: dataByContinent.oneCasePerPeople,
-    },
-    {
-      paperTitle: "one Death Per People",
-      paperAnswer: dataByContinent.oneDeathPerPeople,
-    },
-    {
-      paperTitle: "one Test Per People",
-      paperAnswer: dataByContinent.oneTestPerPeople,
-    },
-    {
-      paperTitle: "population",
-      paperAnswer: dataByContinent.population,
-    },
-    {
-      paperTitle: "recovered",
-      paperAnswer: dataByContinent.recovered,
-    },
-    {
-      paperTitle: "recovered Per One Million",
-      paperAnswer: dataByContinent.recoveredPerOneMillion,
-    },
-    {
-      paperTitle: "tests",
-      paperAnswer: dataByContinent.tests,
-    },
-    {
-      paperTitle: "tests Per One Million",
-      paperAnswer: dataByContinent.testsPerOneMillion,
-    },
-    {
-      paperTitle: "today's Cases",
-      paperAnswer: dataByContinent.todayCases,
-    },
-    {
-      paperTitle: "today's Deaths",
-      paperAnswer: dataByContinent.todayDeaths,
-    },
-    {
-      paperTitle: "today's Recoveries",
-      paperAnswer: dataByContinent.todayRecovered,
-    },
-  ];
-
   const getAllWorldCovidData = async () => {
     await fetch(`https://disease.sh/v3/covid-19/all`)
       .then((response) => response.json())
@@ -224,8 +141,8 @@ const CovidWorld = ({ value, index }) => {
     <>
       {value === index && (
         <>
-          <div class="world_wrapper">
-            <div class="select_options">
+          <div className="world_wrapper">
+            <div className="select_options">
               <FormControl variant="filled">
                 <InputLabel id="demo-simple-select-filled-label">
                   Search...
@@ -253,7 +170,7 @@ const CovidWorld = ({ value, index }) => {
                 </FormHelperText>
               </FormControl>
             </div>
-            <div class="world_head">
+            <div className="world_head">
               {selectOptions === "Get COVID19 World Information" && (
                 <>
                   <WorldPaperInformation
@@ -265,7 +182,7 @@ const CovidWorld = ({ value, index }) => {
               {selectOptions === "Get COVID19 Data by continents" && (
                 <>
                   <ContinentPaperInformation
-                    ContinentsPaperContent={ContinentsPaperContent}
+                    dataByContinent={dataByContinent}
                   />
                 </>
               )}
